LIBILS-176. Moved JAXB unmarshalling classes into "jaxb" subdirectory
Added "jaxb.index" to allow simpler configuration of JAXBContext.

Updated JAXBContext calls to use "edu.umd.lib.axsjw.jaxb" package
instead of listing classes individually.

https://issues.umd.edu/browse/LIBILS-176
