+++ src/main/java/edu/umd/lib/axsjw/jaxb/CircStatus.java
@@ -1,4 +1,4 @@
-package edu.umd.lib.axsjw;
+package edu.umd.lib.axsjw.jaxb;
 
 import java.util.Map;
