+++ src/test/java/edu/umd/lib/axsjw/jaxb/CircStatusTest.java
@@ -1,4 +1,4 @@
-package edu.umd.lib.axsjw;
+package edu.umd.lib.axsjw.jaxb;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
@@ -20,13 +20,10 @@ public class CircStatusTest {
   @Before
   public void setUp() throws Exception {
     // Create context
-    JAXBContext ctx = JAXBContext.newInstance(CircStatus.class);
+    JAXBContext ctx = JAXBContext.newInstance("edu.umd.lib.axsjw.jaxb");
 
     // Create unmarshaller
     unmarshaller = ctx.createUnmarshaller();
-
-    MapAdapter mapAdapter = new MapAdapter();
-    unmarshaller.setAdapter(mapAdapter);
   }
 
   @After
